+++ package.json
@@ -17,7 +17,8 @@
 	"dependencies": {
 		"react": "16.0.0",
 		"react-native": "0.50.3",
-		"react-native-config": "0.10.0"
+		"react-native-config": "0.10.0",
+		"react-router-native": "4.2.0"
 	},
 	"devDependencies": {
 		"babel-eslint": "8.0.2",
