+++ src/modules/home/constants/instructions.js
@@ -1,0 +1,8 @@
+import { Platform } from 'react-native'
+
+export const instructions = Platform.select({
+  ios: 'Press Cmd+R to reload,\n' +
+    'Cmd+D or shake for dev menu',
+  android: 'Double tap R on your keyboard to reload,\n' +
+    'Shake or press menu button for dev menu'
+})
